# shared/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Homebrewery API",
    "description": "The Homebrewery WYSIWYG back end (docs/wysiwyg-plan.md §8). Same origin as the SPA: cookie sign-in, and writes need an Origin header matching the site. Errors are problem+json.",
    "version": "v1"
  },
  "paths": {
...
    "/api/vault": {
      "get": {
        "tags": [
          "Vault"
        ],
        "summary": "Search published brews (title, description and text), sorted and paged.",
        "description": "`q` uses web search syntax with the `simple` text search configuration (no stemming, any language). `sort`: relevance (default with `q`), updated (default without), created, views, title. `dir`: asc or desc (default asc for title, desc otherwise). `pageSize` is clamped to 1-60 (default 20) and `page` to 1-10000. An unknown `sort` or `dir`, or a `q` over 256 characters, is a 400 validation problem.",
        "operationId": "SearchVault",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "enum": [
                "relevance",
                "updated",
                "created",
                "views",
                "title"
              ],
              "type": "string"
            }
          },
          {
            "name": "dir",
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{handle}/brews": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "A user's brews: published ones for everyone; all of them, with edit ids, for the user.",
        "operationId": "ListUserBrews",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBrewList"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/notifications": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Every notification (past, current and future), newest start first.",
        "operationId": "AdminListNotifications",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationInfo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a notification (409 when the dismiss key is taken).",
        "operationId": "AdminCreateNotification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/notifications/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "One notification.",
        "operationId": "AdminGetNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Replace a notification's fields.",
        "operationId": "AdminUpdateNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a notification.",
        "operationId": "AdminDeleteNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthorRole": {
        "enum": [
          "owner",
          "author",
          "invited"
        ]
      },
...
      "BrewSummary": {
        "required": [
          "shareId",
          "editId",
          "title",
          "description",
          "tags",
          "authors",
          "theme",
          "lang",
          "pageCount",
          "views",
          "published",
          "thumbnailUrl",
          "createdAt",
          "updatedAt",
          "lastViewedAt",
          "role",
          "locked"
        ],
        "type": "object",
        "properties": {
          "shareId": {
            "type": "string"
          },
          "editId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Only in the caller's own list (`GET /api/users/{handle}/brews` as that user); otherwise null."
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Handles of the owner and authors in display order (invited users excluded)."
          },
          "theme": {
            "type": "string",
            "description": "A static theme key or a user theme's share id."
          },
          "lang": {
            "type": "string"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "views": {
            "type": "integer",
            "format": "int32"
          },
          "published": {
            "type": "boolean",
            "description": "Always true in the vault and in other people's lists."
          },
          "thumbnailUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastViewedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When the share page last counted a view, or null.",
            "format": "date-time"
          },
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "The caller's role; only in the caller's own list, otherwise null.",
                "$ref": "#/components/schemas/AuthorRole"
              }
            ]
          },
          "locked": {
            "type": "boolean",
            "description": "Whether a moderation lock blocks the share page. Locked brews appear only in their authors'\n    own lists."
          }
        },
        "description": "A brew in a list (the vault, a user's page). Never contains the document."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "NotificationInfo": {
        "required": [
          "id",
          "dismissKey",
          "title",
          "body",
          "startsAt",
          "stopsAt",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dismissKey": {
            "type": "string",
            "description": "Unique key; the client remembers it when the user dismisses the notification."
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string",
            "description": "Plain text (never markdown or HTML)."
          },
          "startsAt": {
            "type": "string",
            "description": "Shown from this time (inclusive).",
            "format": "date-time"
          },
          "stopsAt": {
            "type": "string",
            "description": "Shown until this time (exclusive).",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "A site notification (banner)."
      },
      "NotificationInput": {
        "required": [
          "dismissKey",
          "title",
          "body",
          "startsAt",
          "stopsAt"
        ],
        "type": "object",
        "properties": {
          "dismissKey": {
            "type": [
              "null",
              "string"
            ],
            "description": "Required, unique, at most 100 characters (409 when another notification uses it)."
          },
          "title": {
            "type": [
              "null",
              "string"
            ],
            "description": "Required, at most 200 characters."
          },
          "body": {
            "type": [
              "null",
              "string"
            ],
            "description": "Plain text, at most 10000 characters."
          },
          "startsAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "Default: now.",
            "format": "date-time"
          },
          "stopsAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "Required; after StartsAt.",
            "format": "date-time"
          }
        },
        "description": "Body of `POST /api/admin/notifications` and `PUT /api/admin/notifications/{id}`."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UserBrewList": {
        "required": [
          "handle",
          "own",
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "handle": {
            "type": "string",
            "description": "The user's handle."
          },
          "own": {
            "type": "boolean",
            "description": "True when the caller is this user: the list then includes unpublished and locked brews and brews\n    the user is invited to, with edit ids and roles."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrewSummary"
            },
            "description": "At most 1000 brews. In the user's own list, brews they are only invited to come after the others."
          },
          "total": {
            "type": "integer",
            "description": "All brews the caller may see in this list (may exceed the items returned).",
            "format": "int32"
          }
        },
        "description": "A user's brews (`GET /api/users/{handle}/brews`), most recently updated first."
      },
...
      "VaultPage": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "sort",
          "dir"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrewSummary"
            }
          },
          "total": {
            "type": "integer",
            "description": "Matching brews over all pages.",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "description": "The page returned (after clamping).",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "The page size used (after clamping).",
            "format": "int32"
          },
          "sort": {
            "type": "string",
            "description": "The sort applied (`relevance` without `q` reports `updated`)."
          },
          "dir": {
            "type": "string",
            "description": "The direction applied."
          }
        },
        "description": "A page of vault results (`GET /api/vault`)."
      }
    }
  },
...
}